Extract reusable helper to mock engine with provided Json resource file path
